Crate penumbra_sdk_txhash

Structs

AuthHash

A hash of a transaction’s authorizing data, describing both its effects on the chain state as well as the cryptographic authorization of those effects.

EffectHash

A hash of a transaction’s effecting data, describing its effects on the chain state.

TransactionContext

Stateless verification context for a transaction.

TransactionId

A transaction ID (hash), the Sha256 hash used by Tendermint to identify transactions.

Traits

AuthorizingData

EffectingData

Something that can be hashed to produce an EffectHash.